# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MyAds Dashboard API",
        "description": "API for managing advertising applications with role-based access control (Admin/Publisher)",
        "contact": {
            "name": "MyAds Support",
            "email": "admin@myads.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Local Development Server"
        },
        {
            "url": "http://localhost:5174/api",
            "description": "Frontend Proxy (Vite)"
        }
    ],
    "paths": {
...
        "/apps/{id}/stats": {
            "get": {
                "tags": [
                    "Apps"
                ],
                "summary": "Get app statistics",
                "description": "Retrieve statistics for a specific application",
                "operationId": "7ed722030699581a003161fe93523dd9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Application ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Application statistics",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "total_impressions": {
                                            "type": "integer",
                                            "example": 1500
                                        },
                                        "today_impressions": {
                                            "type": "integer",
                                            "example": 150
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "active"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden - Not the app owner"
                    },
                    "404": {
                        "description": "Application not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication. Format: Bearer {token}",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}